--- src/infra/http/controllers/specialty.controller.e2e-spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { SpecialtyController } from './specialty.controller';
-
-describe('SpecialtyController', () => {
-	let controller: SpecialtyController;
-
-	beforeEach(async () => {
-		const module: TestingModule = await Test.createTestingModule({
-			controllers: [SpecialtyController],
-		}).compile();
-
-		controller = module.get<SpecialtyController>(SpecialtyController);
-	});
-
-	it('should be defined', () => {
-		expect(controller).toBeDefined();
-	});
-});
