+++ src/infra/http/controllers/treatment-type/treatment-type.controller.ts
